Move endpoints.akkahttp.client.JsonSchemaEntities to a separate artifact
Introduce an endpoints-akka-http-client-circe artifact that depends
on Circe, and remove circe dependency from endpoints-akka-http-client
artifact.

Fixes #303
